+++ mirror/storage_test.go
@@ -8,7 +8,7 @@ import (
 	"github.com/cybozu-go/aptutil/apt"
 )
 
-func TestStorageBadConstruction(t *testing.T) {
+func testStorageBadConstruction(t *testing.T) {
 	t.Parallel()
 
 	f, err := ioutil.TempFile("", "gotest")
@@ -34,7 +34,7 @@ func TestStorageBadConstruction(t *testing.T) {
 	}
 }
 
-func TestStorage(t *testing.T) {
+func testStorageLookup(t *testing.T) {
 	t.Parallel()
 
 	d, err := ioutil.TempDir("", "gotest")
@@ -67,7 +67,7 @@ func TestStorage(t *testing.T) {
 	}
 
 	fi := apt.MakeFileInfo("a/b/c", []byte{'a', 'b', 'd'})
-	fi2, fullpath := s.Lookup(fi)
+	fi2, fullpath := s.Lookup(fi, false)
 	if fi2 != nil {
 		t.Error(`fi2 != nil`)
 	}
@@ -75,7 +75,7 @@ func TestStorage(t *testing.T) {
 		t.Error(`len(fullpath) != 0`)
 	}
 
-	fi3, _ := s.Lookup(apt.MakeFileInfo("a/b/c", files["a/b/c"]))
+	fi3, _ := s.Lookup(apt.MakeFileInfo("a/b/c", files["a/b/c"]), false)
 	if fi3 == nil {
 		t.Error(`fi3 == nil`)
 	}
@@ -92,16 +92,81 @@ func TestStorage(t *testing.T) {
 		t.Error(err)
 	}
 
-	fi4, _ := s2.Lookup(apt.MakeFileInfo("a/b/c", files["a/b/c"]))
+	fi4, _ := s2.Lookup(apt.MakeFileInfo("a/b/c", files["a/b/c"]), false)
 	if fi4 == nil {
 		t.Error(`fi4 == nil`)
 	}
-	fi5, _ := s2.Lookup(apt.MakeFileInfo("def", files["def"]))
+	fi5, _ := s2.Lookup(apt.MakeFileInfo("def", files["def"]), false)
 	if fi5 == nil {
 		t.Error(`fi5 == nil`)
 	}
-	fi6, _ := s2.Lookup(apt.MakeFileInfo("a/pp/le", files["a/pp/le"]))
+	fi6, _ := s2.Lookup(apt.MakeFileInfo("a/pp/le", files["a/pp/le"]), false)
 	if fi6 == nil {
 		t.Error(`fi6 == nil`)
 	}
+	fi7, _ := s2.Lookup(apt.MakeFileInfo("a/pp/le", files["def"]), false)
+	if fi7 != nil {
+		t.Error(`fi7 != nil`)
+	}
+}
+
+func testStorageStore(t *testing.T) {
+	t.Parallel()
+
+	d, err := ioutil.TempDir("", "gotest")
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer os.RemoveAll(d)
+
+	s, err := NewStorage(d, "pre")
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	err = s.Load()
+	if err != nil {
+		t.Error(err)
+	}
+
+	fn := "a/b/c"
+	data := []byte{'a', 'b', 'c'}
+	fi := apt.MakeFileInfo(fn, data)
+
+	err = s.Store(fi, data)
+	if err != nil {
+		t.Error(err)
+	}
+	found, _ := s.Lookup(fi, false)
+	if found == nil {
+		t.Error(`found == nil`)
+	}
+
+	// duplicates should not be granted
+	err = s.Store(fi, data)
+	if err == nil {
+		t.Error(`err == nil`)
+	}
+
+	data2 := []byte{'d', 'e', 'f'}
+	fi2 := apt.MakeFileInfo(fn, data2)
+
+	err = s.StoreWithHash(fi2, data2)
+	if err != nil {
+		t.Error(err)
+	}
+	notfound, _ := s.Lookup(fi2, false)
+	if notfound != nil {
+		t.Error(`notfound != nil`)
+	}
+	found, _ = s.Lookup(fi2, true)
+	if found == nil {
+		t.Error(`found == nil`, d, fi2.SHA256Path())
+	}
+}
+
+func TestStorage(t *testing.T) {
+	t.Run("BadConstruction", testStorageBadConstruction)
+	t.Run("Lookup", testStorageLookup)
+	t.Run("Store", testStorageStore)
 }